% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean_Assets.R
\name{clean_assets}
\alias{clean_assets}
\title{Clean Assets data}
\usage{
clean_assets(data, main_data, benunit_data)
}
\arguments{
\item{data}{Data table. The combined adult, child, and household raw FRS data files.}

\item{main_data}{Data table. The main data file "frsxxyy"}

\item{benunit_data}{Data table. The raw FRS benefit unit data file.}
}
\value{
Data table
}
\description{
Generate clean variables for assets from the raw
Family Resources Survey data sets.
}
\examples{


\dontrun{

}
}
