% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean_Benefits.R
\name{clean_benefits}
\alias{clean_benefits}
\title{Clean Income and Tax data}
\usage{
clean_benefits(data, benefits_data, benunit_data)
}
\arguments{
\item{data}{Data table. The combined adult, child, and household raw FRS data files.}

\item{benefits_data}{Data table. The raw FRS benefits file.}

\item{benunit_data}{Data table. The raw FRS benefit unit file.}
}
\value{
Data table
}
\description{
Generate clean variables for individual market income (excluding benefit income) and taxes from the raw
Family Resources Survey data sets that are needed for the microsimulation model.
}
\examples{


\dontrun{

}
}
