% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean_Demographic.R
\name{clean_demographic}
\alias{clean_demographic}
\title{Clean Demographic data}
\usage{
clean_demographic(data, main_data)
}
\arguments{
\item{data}{Data table. The combined adult, child, and household raw FRS data files.}

\item{main_data}{Data table. The main data file "frsxxyy"}
}
\description{
Generate clean variables for individual demographics and generate identifiers from the raw
Family Resources Survey data sets that are needed for the microsimulation model.
}
\examples{


\dontrun{

}
}
