% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean_Expenditure.R
\name{clean_expenditure}
\alias{clean_expenditure}
\title{Clean Expenditures data}
\usage{
clean_expenditure(data, maint_data, penprov_data)
}
\arguments{
\item{data}{Data table. The combined adult, child, and household raw FRS data files.}

\item{maint_data}{Data table. The raw FRS maintenance data file.}

\item{penprov_data}{Data table. The raw FRS pension provision data file.}
}
\description{
Generate clean variables for individual expenditures from the raw
Family Resources Survey data sets that are needed for the microsimulation model.
}
\examples{


\dontrun{

}
}
