% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean_Health.R
\name{clean_health}
\alias{clean_health}
\title{Clean Health data}
\usage{
clean_health(data)
}
\arguments{
\item{data}{Data table. The combined adult, child, and household raw FRS data files.}

\item{main_data}{Data table. The main data file "frsxxyy"}
}
\value{
Data table
}
\description{
Generate clean variables for individual status.
}
\examples{


\dontrun{

}
}
