% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean_Income.R
\name{clean_income}
\alias{clean_income}
\title{Clean Income and Tax data}
\usage{
clean_income(
  data,
  pension_data,
  job_data,
  oddjob_data,
  accounts_data,
  benefits_data
)
}
\arguments{
\item{data}{Data table. The combined adult, child, and household raw FRS data files.}

\item{pension_data}{Data table. The raw FRS pension data file.}

\item{job_data}{Data table. The raw FRS job file.}

\item{oddjob_data}{Data table. The raw FRS odd-job file.}

\item{accounts_data}{Data table. the raw FRS accounts file.}

\item{benefits_data}{Data table. The raw FRS benefits file.}
}
\value{

}
\description{
Generate clean variables for individual market income (excluding benefit income) and taxes from the raw
Family Resources Survey data sets that are needed for the microsimulation model.
}
\examples{


\dontrun{

}
}
