% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean_LabMarket.R
\name{clean_labmarket}
\alias{clean_labmarket}
\title{Clean Labour Market data}
\usage{
clean_labmarket(data, main_data, job_data, income_data, year)
}
\arguments{
\item{data}{Data table. The combined adult, child, and household raw FRS data files.}

\item{main_data}{Data table. The main data file "frsxxyy"}

\item{job_data}{Data table. The raw FRS job file.}

\item{income_data}{Data table. The income data cleaned by the `clean_income` function.}

\item{year}{Numeric integer - year corresponding to the start of the financial year i.e. 2020/21 data is indexed as 2020}
}
\value{
Data table
}
\description{
Generate clean variables for labour market information from the raw
Family Resources Survey data sets.
}
\details{
Note that this function requires data generated from the `clean_income` function using the
same input data
}
\examples{


\dontrun{

}
}
