% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frsclean.R
\name{frsclean}
\alias{frsclean}
\title{Read, Clean, and Combine Family Resources Survey data}
\usage{
frsclean(
  root,
  file,
  years = 2020,
  ages = NULL,
  keep_vars = NULL,
  complete_vars = NULL
)
}
\arguments{
\item{root}{Character - the root directory.}

\item{file}{Character - the file path and name.}

\item{years}{Integer vector - the years of the FRS to read, clean, and combine (defaults to 2020).}

\item{ages}{Integer vector - the ages in single years to retain. If NULL (default), retain all ages.}

\item{keep_vars}{Character vector - the names of the variables to keep (defaults to NULL - retaining all variables).}

\item{complete_vars}{Character vector - the names of the variables on which the selection of complete cases will be based.
If NULL (default) no complete-case filtering is applied.}
}
\value{
Returns a new set of variables
}
\description{
A wrapper function for applying all of the reading and cleaning functions for the
Family Resources Survey (FRS) end user licence data, selecting the
desired variables/observations for the analysis, and specifying complete case
restrictions.
}
