% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inflation}
\alias{inflation}
\title{Annual Inflation}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 70 rows and 3 columns.
}
\usage{
inflation
}
\description{
Time series of the CPIH (consumer price index including housing costs) inflation index and the
RPI (retail price index) on an annual basis with a base year of 2022. Data are obtained
from the Office for National Statistics for \href{https://www.ons.gov.uk/generator?format=xls&uri=/economy/inflationandpriceindices/timeseries/l522/mm23}{CPIH}
and \href{https://www.ons.gov.uk/generator?format=xls&uri=/economy/inflationandpriceindices/timeseries/chaw/mm23}{RPI}
last accessed 16/06/2023.
}
\keyword{datasets}
