% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_frs_2019_20.R
\name{read_frs_2019_20}
\alias{read_frs_2019_20}
\title{Read Family Resources Survey 2019/2020}
\usage{
read_frs_2019_20(
  root = "X:/",
  file =
    "HAR_PR/PR/IPS_beyond_SMI_NIHR202996/General/R/data-family-resources-survey/data/raw"
)
}
\arguments{
\item{root}{Character. The root directory}

\item{file}{Character. The file path and name}
}
\value{
Data table
}
\description{
Read in the Family Resources Survey data for the 2019-20 fiscal year.
}
\examples{


\dontrun{


}
}
