% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_frs_2020.R
\name{read_frs_2020}
\alias{read_frs_2020}
\title{Read Family Resources Survey 2020}
\usage{
read_frs_2020(
  root = "C:/",
  file = "Users/damon/OneDrive/Documents/Datasets/Family Resources Survey/tab"
)
}
\arguments{
\item{root}{Character. The root directory}

\item{file}{Character. The file path and name}
}
\value{

}
\description{
Read in the Family Resources Survey data for the 2020-21 fiscal year, including the
two years prior if there is a need for a larger sample size
}
\examples{


\dontrun{

}
}
