% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_frs_2021_22.R
\name{read_frs_2021_22}
\alias{read_frs_2021_22}
\title{Read Family Resources Survey 2021/2022}
\usage{
read_frs_2021_22(
  root = "X:/",
  file =
    "HAR_PR/PR/IPS_beyond_SMI_NIHR202996/General/R/data-family-resources-survey/data/raw"
)
}
\arguments{
\item{root}{Character. The root directory}

\item{file}{Character. The file path and name}
}
\value{
Data table
}
\description{
Read in the Family Resources Survey data for the 2021-22 fiscal year, including the
two years prior if there is a need for a larger sample size
}
\examples{


\dontrun{


}
}
