% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_frs_2022_23.R
\name{read_frs_2022_23}
\alias{read_frs_2022_23}
\title{Read Family Resources Survey 2022/2023}
\usage{
read_frs_2022_23(
  root = "X:/",
  file =
    "HAR_PR/PR/IPS_beyond_SMI_NIHR202996/General/R/data-family-resources-survey/data/raw"
)
}
\arguments{
\item{root}{Character. The root directory}

\item{file}{Character. The file path and name}
}
\value{
Data table
}
\description{
Read in the Family Resources Survey data for the 2022-23 fiscal year
}
\examples{


\dontrun{


}
}
