% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_data.R
\name{select_data}
\alias{select_data}
\title{Select variables and apply filters}
\usage{
select_data(data, ages = NULL, keep_vars = NULL, complete_vars = NULL)
}
\arguments{
\item{data}{Data table. the FRS dataset.}

\item{ages}{Integer vector - the ages in single years to retain. If NULL (default), retain all ages.}

\item{keep_vars}{Character vector - the names of the variables to keep (defaults to NULL - retaining all variables).}

\item{complete_vars}{Character vector - the names of the variables on which the selection of complete cases will be based.
If NULL (default) no complete-case filtering is applied.}
}
\value{
Returns a reduced version of data
}
\description{
Selects the variables required for analysis and selects only the rows without missing data
for specified variables.
}
\examples{

\dontrun{
#'
}

}
