% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frs_clean_global.R
\name{frs_clean_global}
\alias{frs_clean_global}
\title{Apply Cleaning Modules}
\usage{
frs_clean_global(
  data_list,
  ages = NULL,
  keep_vars = NULL,
  complete_vars = NULL
)
}
\arguments{
\item{data_list}{List. List of data tables returned from the FRS cleaning functions.}

\item{ages}{Integer vector - the ages in single years to retain. If NULL (default), retain all ages.}

\item{keep_vars}{Character vector - the names of the variables to keep (defaults to NULL - retaining all variables).}

\item{complete_vars}{Character vector - the names of the variables on which the selection of complete cases will be based.
If NULL (default) no complete-case filtering is applied.}
}
\value{
Returns a new set of variables
}
\description{
A wrapper function for applying all of the cleaning functions and selecting the
desired variables/observations for the analysis by applying the select_data function.
}
