% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inflation_Adjust.R
\name{Inflation_Adjust}
\alias{Inflation_Adjust}
\title{Inflation Adjustment}
\usage{
Inflation_Adjust(data, year = NULL, index = "cpih")
}
\arguments{
\item{data}{Data table - the FRS dataset.}

\item{year}{Numeric integer - year corresponding to the start of the financial year i.e. 2020/21 data is indexed as 2020}

\item{index}{Character - inflation index to use for real terms adjustment, "cpih" (default) or "rpi"}
}
\value{
Data table
}
\description{
Convert all monetary variables into real terms
}
\examples{


\dontrun{

}
}
